backend/donna/tools/social.py
"""
Social media integration tools for Donna.

Handles TikTok and Instagram analytics (where APIs are available).
"""

from typing import Optional
from datetime import datetime

from langchain_core.tools import tool

from donna.config import get_settings


@tool
def get_tiktok_analytics() -> str:
    """
    Get TikTok analytics.
    
    Note: TikTok's API has limited access. This is a placeholder
    for when you have API access.
    
    Returns basic account stats and recent video performance.
    """
    settings = get_settings()
    
    # TikTok API requires business account and approval
    return """⚠️ TikTok API Not Configured

TikTok's API requires:
1. A TikTok Business Account
2. Developer application approval
3. API credentials

For now, you can:
- Add content ideas via `/idea tiktok [description]`
- Track ideas in `content/tiktok/ideas.md`

Apply for API access at: https://developers.tiktok.com/
"""


@tool
def get_instagram_analytics() -> str:
    """
    Get Instagram analytics.
    
    Note: Requires Instagram Business/Creator account and
    Facebook Graph API access.
    
    Returns account stats and recent post performance.
    """
    settings = get_settings()
    
    # Instagram API requires Facebook Graph API + business account
    return """⚠️ Instagram API Not Configured

Instagram's API requires:
1. An Instagram Business or Creator account
2. A connected Facebook Page
3. Facebook Graph API credentials

For now, you can:
- Add content ideas via `/idea instagram [description]`
- Track ideas in `content/instagram/ideas.md`

Set up at: https://developers.facebook.com/
"""


@tool
def plan_content_calendar(
    platform: str,
    start_date: Optional[str] = None,
    days: int = 7
) -> str:
    """
    Generate a content calendar suggestion.
    
    Args:
        platform: "youtube", "tiktok", "instagram", or "all"
        start_date: Start date in YYYY-MM-DD format (defaults to today)
        days: Number of days to plan
    
    Returns a suggested posting schedule based on best practices.
    """
    from datetime import timedelta
    
    if start_date:
        start = datetime.strptime(start_date, "%Y-%m-%d")
    else:
        start = datetime.now()
    
    platform = platform.lower()
    
    # Platform-specific posting recommendations
    schedules = {
        "youtube": {
            "frequency": "1-2 per week",
            "best_days": ["Tuesday", "Thursday", "Saturday"],
            "best_time": "2-4 PM",
            "content_types": ["Long-form tutorials", "Behind the scenes", "Project updates"]
        },
        "tiktok": {
            "frequency": "1-3 per day",
            "best_days": ["Daily"],
            "best_time": "9 AM, 12 PM, 7 PM",
            "content_types": ["Quick tips", "Day in life", "Build process", "Trending sounds"]
        },
        "instagram": {
            "frequency": "1 post + 3-5 stories daily",
            "best_days": ["Tuesday", "Wednesday", "Friday"],
            "best_time": "11 AM - 1 PM, 7 PM",
            "content_types": ["Carousels", "Reels", "Stories", "Behind the scenes"]
        }
    }
    
    lines = [f"# Content Calendar: {start.strftime('%B %d')} - {(start + timedelta(days=days-1)).strftime('%B %d')}\n"]
    
    platforms_to_show = [platform] if platform != "all" else ["youtube", "tiktok", "instagram"]
    
    for p in platforms_to_show:
        if p not in schedules:
            continue
        
        sched = schedules[p]
        lines.append(f"## {p.title()}")
        lines.append(f"**Recommended Frequency**: {sched['frequency']}")
        lines.append(f"**Best Days**: {', '.join(sched['best_days'])}")
        lines.append(f"**Best Times**: {sched['best_time']}")
        lines.append("")
        lines.append("**Content Ideas**:")
        for ct in sched['content_types']:
            lines.append(f"- {ct}")
        lines.append("")
        
        # Generate specific slots
        lines.append("**Suggested Posts**:")
        for i in range(days):
            date = start + timedelta(days=i)
            day_name = date.strftime("%A")
            
            if day_name in sched['best_days'] or "Daily" in sched['best_days']:
                lines.append(f"- {date.strftime('%a %m/%d')}: [Plan content here]")
        
        lines.append("")
    
    lines.append("---")
    lines.append("Use `/idea [platform] [description]` to add specific ideas.")
    
    return "\n".join(lines)


