backend/pyproject.toml
[project]
name = "donna"
version = "1.0.0"
description = "Donna - Personal AI Executive Assistant"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Dallion King"}
]

dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "langgraph>=0.0.40",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "openai>=1.10.0",
    "tiktoken>=0.5.2",
    "supabase>=2.3.0",
    "vecs>=0.4.0",
    "google-api-python-client>=2.111.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "python-telegram-bot>=20.7",
    "apscheduler>=3.10.4",
    "croniter>=2.0.1",
    "httpx>=0.26.0",
    "aiohttp>=3.9.0",
    "PyGithub>=2.1.1",
    "python-dateutil>=2.8.2",
    "pytz>=2024.1",
    "structlog>=24.1.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]


